Changes necessary to super peer network (1 for cleanup, 2 to make Laith's new signaling server work)
