+++ matchbox_super_client/Cargo.toml
@@ -3,7 +3,7 @@ name = "matchbox_super_client"
 version = "0.1.0"
 edition = "2021"
 
-[dependencies]                                
+[dependencies]
 wasm-bindgen = {version = "0.2", features = ["serde-serialize"]}     
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
@@ -23,8 +23,5 @@ features = [
     "v4", "js"
 ]
 
-[dev-dependencies]
-wasm-bindgen-test = "0.2"
-
 [lib]
-crate-type = ["cdylib"]
+crate-type = ["cdylib", "rlib"]
